efficientnet/requirements.txt
numpy>=1.26.0
pandas>=2.1.0
torch>=2.2.0
torchvision>=0.17.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
pillow>=10.2.0
tqdm>=4.66.0
efficientnet-pytorch>=0.7.1
tensorboard>=2.15.0
seaborn>=0.13.0 